refactor(discord): big cleanup of the architecture
