+++ src/discord/channels/functions/is-discord-text-channel.ts
