+++ src/discord/mentions/functions/is-discord-message-mentions.ts
