+++ src/discord/users/functions/is-discord-client-user.ts
