+++ src/discord/users/functions/is-discord-user.ts
